Repair load when file length grows during read
