+++ src/libraries/filesystem/src/CMakeLists.txt
@@ -7,13 +7,8 @@ target_sources(m_filesystem PRIVATE
     path_conversions.cpp
     seekable_input_file.cpp
     seekable_output_file.cpp
- "monitor.cpp")
-
-if(WIN32)
-    target_sources(m_filesystem PRIVATE
-        windows_loadstore.cpp 
-    )
-endif()
+    monitor.cpp
+)
 
 target_compile_definitions(m_filesystem PRIVATE _CRT_SECURE_NO_WARNINGS)
