+++ src/libraries/filesystem/src/platforms/windows/CMakeLists.txt
@@ -3,6 +3,7 @@ cmake_minimum_required(VERSION 3.23)
 target_sources(m_filesystem PRIVATE
     directory_watcher.cpp
     string_to_path_native.cpp
+    windows_loadstore.cpp
 )
 
 target_link_libraries(m_filesystem PUBLIC
